refactor(core): extract components from `index` files
