+++ .babelrc
@@ -6,6 +6,7 @@
       {
         "root": ["./"]
       }
-    ]
+    ],
+    "@babel/plugin-proposal-export-default-from"
   ]
 }
