+++ package.json
@@ -83,6 +83,7 @@
   "homepage": "https://github.com/babajka/babajka-frontend#readme",
   "main": "index.js",
   "devDependencies": {
+    "@babel/plugin-proposal-export-default-from": "^7.0.0-beta.54",
     "@zeit/next-bundle-analyzer": "^0.1.1",
     "babel-eslint": "^8.0.2",
     "babel-plugin-module-resolver": "^3.1.1",
